requirements.txt
Pillow>=10.0.0
reportlab>=4.0.0
pywin32>=306
pyinstaller>=6.0.0

